refactor: extract OpenAPI builder and improve docs for text/html
- Remove Zodios dependency and related OpenAPI spec generation code
- Integrate new function to create OpenAPI specification
- Maintain existing functionality for serving OpenAPI spec as JSON
